requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.6.0
click==8.1.3
colorama==0.4.5
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8==4.0.1
flake8-html==0.4.2
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
